requirements.txt
aiogram
python-dotenv
sqlalchemy
asyncpg
fastapi
uvicorn
sqladmin